refactor: search_quality 패키지를 test 패키지 아래로 이동
