--- src/main/resources/ground-truth.json
@@ -1,273 +1,0 @@
-[
-  {
-    "query": "Java",
-    "idealResultsMap": {
-      "103": 5,
-      "147": 4,
-      "50": 3,
-      "294": 2,
-      "494": 1
-    }
-  },
-  {
-    "query": "TypeScript",
-    "idealResultsMap": {
-      "327": 5,
-      "518": 4,
-      "997": 3,
-      "452": 2,
-      "25": 1
-    }
-  },
-  {
-    "query": "Kotlin",
-    "idealResultsMap": {
-      "147": 5,
-      "494": 4,
-      "50": 3,
-      "441": 2,
-      "366": 1
-    }
-  },
-  {
-    "query": "Go",
-    "idealResultsMap": {
-      "232": 5
-    }
-  },
-  {
-    "query": "Rust",
-    "idealResultsMap": {
-      "442": 5,
-      "98": 4,
-      "978": 3,
-      "75": 2,
-      "70": 1
-    }
-  },
-  {
-    "query": "C++",
-    "idealResultsMap": {
-      "431": 5,
-      "88": 4,
-      "89": 3,
-      "429": 2,
-      "438": 1
-    }
-  },
-  {
-    "query": "Swift",
-    "idealResultsMap": {
-      "437": 5,
-      "135": 4,
-      "510": 3,
-      "192": 2,
-      "495": 1
-    }
-  },
-  {
-    "query": "Scala",
-    "idealResultsMap": {
-      "442": 5,
-      "408": 4,
-      "435": 3,
-      "411": 2,
-      "438": 1
-    }
-  },
-  {
-    "query": "Spring",
-    "idealResultsMap": {
-      "83": 5,
-      "105": 4,
-      "509": 3,
-      "234": 2,
-      "529": 1
-    }
-  },
-  {
-    "query": "Spring Boot",
-    "idealResultsMap": {
-      "105": 5,
-      "83": 4,
-      "509": 3,
-      "28": 2,
-      "234": 1
-    }
-  },
-  {
-    "query": "Django",
-    "idealResultsMap": {}
-  },
-  {
-    "query": "Flask",
-    "idealResultsMap": {
-      "507": 5
-    }
-  },
-  {
-    "query": "React",
-    "idealResultsMap": {
-      "79": 5,
-      "62": 4,
-      "85": 3,
-      "349": 2,
-      "41": 1
-    }
-  },
-  {
-    "query": "Vue.js",
-    "idealResultsMap": {
-      "529": 5,
-      "523": 4,
-      "520": 3,
-      "527": 2,
-      "168": 1
-    }
-  },
-  {
-    "query": "Angular",
-    "idealResultsMap": {
-      "349": 5,
-      "101": 4,
-      "112": 3,
-      "121": 2
-    }
-  },
-  {
-    "query": "Next.js",
-    "idealResultsMap": {
-      "168": 5,
-      "219": 4,
-      "170": 3,
-      "511": 2,
-      "375": 1
-    }
-  },
-  {
-    "query": "Nest.js",
-    "idealResultsMap": {
-      "55": 5,
-      "520": 4,
-      "168": 3,
-      "466": 2,
-      "64": 1
-    }
-  },
-  {
-    "query": "FastAPI",
-    "idealResultsMap": {
-      "995": 5,
-      "398": 4,
-      "989": 3
-    }
-  },
-  {
-    "query": "MySQL",
-    "idealResultsMap": {
-      "971": 5,
-      "348": 4,
-      "8": 3,
-      "153": 2,
-      "479": 1
-    }
-  },
-  {
-    "query": "PostgreSQL",
-    "idealResultsMap": {
-      "985": 5,
-      "433": 4,
-      "983": 3,
-      "367": 2,
-      "1004": 1
-    }
-  },
-  {
-    "query": "MongoDB",
-    "idealResultsMap": {}
-  },
-  {
-    "query": "Redis",
-    "idealResultsMap": {
-      "474": 5,
-      "967": 4,
-      "211": 3,
-      "152": 2,
-      "479": 1
-    }
-  },
-  {
-    "query": "Elasticsearch",
-    "idealResultsMap": {
-      "137": 5,
-      "28": 4,
-      "413": 3,
-      "174": 2,
-      "491": 1
-    }
-  },
-  {
-    "query": "DynamoDB",
-    "idealResultsMap": {
-      "519": 5,
-      "376": 4,
-      "433": 3,
-      "979": 2,
-      "981": 1
-    }
-  },
-  {
-    "query": "Cassandra",
-    "idealResultsMap": {
-      "947": 5,
-      "179": 4,
-      "319": 3
-    }
-  },
-  {
-    "query": "Oracle",
-    "idealResultsMap": {
-      "179": 5,
-      "142": 4,
-      "183": 3,
-      "207": 2,
-      "200": 1
-    }
-  },
-  {
-    "query": "MariaDB",
-    "idealResultsMap": {
-      "1004": 5
-    }
-  },
-  {
-    "query": "RDS",
-    "idealResultsMap": {
-      "969": 5,
-      "46": 4,
-      "98": 3,
-      "985": 2,
-      "952": 1
-    }
-  },
-  {
-    "query": "AWS",
-    "idealResultsMap": {
-      "194": 5,
-      "222": 4,
-      "205": 3,
-      "67": 2,
-      "378": 1
-    }
-  },
-  {
-    "query": "Docker",
-    "idealResultsMap": {
-      "463": 5,
-      "395": 4,
-      "243": 3,
-      "54": 2,
-      "66": 1
-    }
-  }
-]
