+++ src/test/java/com/techfork/evaluation/search/GroundTruthItem.java
@@ -1,11 +1,9 @@
-package com.techfork.domain.search_quality;
+package com.techfork.evaluation.search;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 @Data
 @NoArgsConstructor
