+++ src/test/java/com/techfork/evaluation/search/SearchQualityService.java
@@ -1,13 +1,10 @@
-package com.techfork.domain.search_quality;
-
-import org.springframework.stereotype.Service;
+package com.techfork.evaluation.search;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-@Service
 public class SearchQualityService {
 
     /**
